+++ Serilog.Sinks.Network.Test/Serilog.Sinks.Network.Test.csproj
@@ -4,11 +4,10 @@
     <TargetFramework>netcoreapp1.1</TargetFramework>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="xunit" Version="2.2.0"/>
+    <PackageReference Include="xunit" Version="2.3.0-beta2-build3683"/>
     <PackageReference Include="Newtonsoft.Json" Version="10.0.3"/>
     <PackageReference Include="FluentAssertions" Version="4.19.3"/>
-    <PackageReference Include="Microsoft.NETCore.TestHost" Version="1.1.2"/>
-    <PackageReference Include="Microsoft.TestPlatform.TestHost" Version="15.3.0"/>
+    <DotNetCliToolReference Include="dotnet-xunit" Version="2.3.0-beta2-build3683" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Serilog.Sinks.Network\Serilog.Sinks.Network.csproj"/>
